# libidset: add idset_equal(3), improve efficiency of idset_count(3)