--- a/src/common/libidset/idset.c
+++ b/src/common/libidset/idset.c
@@ -263,6 +263,31 @@ size_t idset_count (const struct idset *idset)
     return idset->count;
 }
 
+bool idset_equal (const struct idset *idset1,
+                  const struct idset *idset2)
+{
+    unsigned int id;
+
+    if (!idset1 || !idset2)
+        return false;
+    if (idset_count (idset1) != idset_count (idset2))
+        return false;
+
+    id = vebsucc (idset1->T, 0);
+    while (id < idset1->T.M) {
+        if (vebsucc (idset2->T, id) != id)
+            return false; // id in idset1 not set in idset2
+        id = vebsucc (idset1->T, id + 1);
+    }
+    id = vebsucc (idset2->T, 0);
+    while (id < idset2->T.M) {
+        if (vebsucc (idset1->T, id) != id)
+            return false; // id in idset2 not set in idset1
+        id = vebsucc (idset2->T, id + 1);
+    }
+    return true;
+}
+
 /*
  * vi:tabstop=4 shiftwidth=4 expandtab
  */