--- a/src/common/libidset/idset_private.h
+++ b/src/common/libidset/idset_private.h
@@ -17,6 +17,7 @@
 #include "idset.h"
 
 struct idset {
+    size_t count;
     Veb T;
     int flags;
 };