As a sanity check I was running a test workload to exercise the ingest->job-manager<->scheduler logic, and while doing some profiling I noticed the workload was spending 10% of the total CPU cycles in `vebsucc()`. As it turns out `sched-simple` is a heavy user of `idset_count(3)` (to sort nodes by least or most available), and currently that function needs to traverse the entire vEB tree for every call.

This PR adds a `count` member to `struct idset` and keeps the counted updated with `idset_set` and `idset_clear` calls.

This small change had a large improvement, so at the risk of premature optimization, I went ahead and proposed it here.

On current master, a set of 4K 1 core jobs submitted to a instance with 4K cores across 128 nodes shows
```
sched-bench: starting test with 4096 jobs. broker.pid=14085
sched-bench: ingested 4096 jobs in 16.13s (253.94 job/s)
sched-bench: allocated 4096 jobs in 33.526s (122.17 job/s)
sched-bench: elapsed time: 34.024s (120.38 job/s)
```

With the change to `idset_count(3)`:
```
sched-bench: starting test with 4096 jobs. broker.pid=23102
sched-bench: ingested 4096 jobs in 16.34s (250.67 job/s)
sched-bench: allocated 4096 jobs in 20.115s (203.63 job/s)
sched-bench: elapsed time: 20.545s (199.37 job/s)
```

Where the `ingested` time is the time for `t/ingest/submitbench -r 4096 -f 16` to complete, the `allocated` time is the difference in time between the `submit` event for the first job and the `alloc` event for the last job, and the `elapsed` time is the total time from the start of `submitbench` and all jobs getting allocated resources. (ingest and "scheduling" are happening together here)

While adding new tests, I noticed that there was already an idset comparator function within the  unit test, and since this was needed for sched-simple, I moved  that code into `idset_equal(3)` now exported by libidset.

Man page entry for `idset_equal` is also added, but I can't add it to the manpage stubs for `idset_create.adoc` because of the old problem where `a2x` won't generate more than 9 stubs for a manpage (any workaround for that?)