@garlick 2019-03-02
This looks great!

There are some other man pages that have missing stubs as well, so fixing/working around that would be nice.  I don't really have any ideas on how to address it, though maybe encode/decode could be split out to their own adoc?