@grondo 2019-03-02
Hm: Looks like `asciidoc` is EOL in 2020, with `asciidoctor` or `asciidoc-py3` as suggested replacement:

https://github.com/asciidoc/asciidoc

Until then (or until wider availability of asciidoctor >= 1.5.7), though, splitting .adoc files is the only thing I can think of as well.