@garlick 2019-03-02
IMHO I think this can go in with missing manpage stubs.  There are a few others that are missing stubs I believe, so when we rid ourselves of asciidoc/a2x we can do an audit and fill them all in.