@grondo 2019-03-02
Ok (I do need to reword or update spelling dictionary).

I was going down the route of trying to make a `libidset.adoc.in` that generated identical `idset_create.adoc` and `idset_encode.adoc` each with different stubs listed (but same content). Would be nice *not* to have to do that..