@grondo 2019-03-02
Ok, just forced an update that adds a note to `doc/man3/Makefile.am` to note why `idset_equal.3` is not generated, and slightly rewords documentation so that a spelling dict update isn't needed.